pyproject.toml
[tool.poetry]
name = "monarch-ingest"
version = "0.8.0"
description = ""
authors = [
    "The Monarch Initiative <info@monarchinitiative.org>",
    "Kevin Schaper <kevinschaper@gmail.com>",
    "Victoria Soesanto <victoria.soesanto@cuanschutz.edu>",
    "Glass Elsarboukh <g.elsarboukh@gmail.com>",
    "Kent Shefchek <kent@tislab.org>",
    "Richard Bruskiewich <richard.bruskiewich@delphinai.com>",
]
packages = [
    { include = "monarch_ingest", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
biolink-model = "4.2.6rc3"
bmt = "^1.0.15"
cat-merge = "1.0.2"
kghub-downloader = "^0.4.1"
kgx = "2.6.0"
linkml = "^1.7.8"
linkml-solr = "0.2.0"
multi-indexer = "0.0.5"
# Other Dependencies
botocore = "^1.31"
importlib-metadata = ">=4.6.1"
loguru = "*"
pydantic = "^2.5"
sh = "^1.14.3"
typer = "^0.12"
yamllint = "^1.35.1"
linkml-runtime = "^1.7.5"
# Remove this once cat-merge fixes its pandas dependency
pandas = "2.0.3"
closurizer = "^0.8.0"
notebook = "^7.3.2"
duckdb = "^1.3.0"
pystow = "^0.5.4"
#koza = "^2.0.1"
koza = "^2.0.1"
[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
mkdocs = "^1.4"
mkdocs-material = ">=9.5"
black = "^24.3"
ruff = "*"

[tool.poetry.scripts]
ingest = "monarch_ingest.main:typer_app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.ruff]
line-length = 120
ignore = [
    "F541", # f-strings with no placeholders

]
[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["F811"] # redefinition of unused imports (mock_koza)
